# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FewohBee REST API
  version: '1.0'
  description: |
    Read-only REST API of the FewohBee property management system.

    Authentication uses personal access tokens created in the user profile
    ("API access"). A token always starts with `fwb_` and can be sent either as
    `Authorization: Bearer <token>` or — for clients that only support
    username/password, such as calendar applications — as the password in HTTP
    Basic auth together with the token owner's username. The actual login
    password is never accepted by the API.

    Every token carries scopes. A scope only takes effect when the token owner
    also holds the underlying application role, so a token can never grant more
    than its owner has.

    | Scope | Required role |
    |---|---|
    | `reservations:read` | ROLE_RESERVATIONS_RO |
    | `calendar:read` | ROLE_RESERVATIONS_RO |
    | `statistics:read` | ROLE_STATISTICS |
    | `invoices:read` | ROLE_INVOICES |
    | `prices:read` | ROLE_RESERVATIONS_RO |
    | `tourist-tax:read` | ROLE_OPERATIONS |

    Requests are rate limited to 300 per 5 minutes per token.
  license:
    name: AGPL-3.0
    url: https://github.com/developeregrem/fewohbee/blob/master/LICENSE
servers:
  - url: '{baseUrl}/api/v1'
    description: Your FewohBee installation
    variables:
      baseUrl:
        default: https://your-installation
        description: >-
          Base URL of the installation, including a subdirectory if the app is
...
security:
  - bearerAuth: []
  - basicAuth: []
...
paths:
...
  /prices:
    get:
      tags: [Prices]
      operationId: listPrices
      summary: The configured price catalogue
      description: |
        Requires the `prices:read` scope.

        Reference data: which price rows exist and when they apply. Working out what a
        stay costs from these fields means reimplementing period priority, weekday
        matching, minStay selection and the guest-category modifiers, and the result
        will drift from the invoice. Use `/prices/quote` for amounts.
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [apartment, misc] }
          description: Apartment (room) prices or miscellaneous prices; both when omitted.
        - name: roomCategoryId
          in: query
          schema: { type: integer }
        - name: originId
          in: query
          description: One or more reservation origin ids (repeatable or comma separated).
          schema: { type: string }
        - name: active
          in: query
          schema: { type: string, enum: ['true', 'false', 'all'], default: 'true' }
        - name: bookableOnline
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: The matching price rows.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Price' }
                  meta:
                    type: object
                    properties:
                      type: { type: [string, 'null'] }
                      roomCategoryId: { type: [integer, 'null'] }
                      originIds: { type: [array, 'null'], items: { type: integer } }
                      active: { type: [boolean, 'null'] }
                      bookableOnline: { type: [boolean, 'null'] }
                      count: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /prices/quote:
    get:
      tags: [Prices]
      operationId: getPriceQuote
      summary: What a stay would cost
      description: |
        Requires the `prices:read` scope.

        Computed through the same pipeline that produces the invoice, so a quote and the
        later invoice agree as long as nobody edits the invoice by hand. A quote is never
        a document — the authoritative amount for a billed stay is the invoice itself.

        `priceFound: false` means no price row applies to this apartment, occupancy and
        period. That is different from a stay that genuinely costs nothing; all totals
        are `0.0` in that case, so never read a `0.00` total as "free".

        `touristTax` and `grandTotal` are `null` when the token lacks the
        `tourist-tax:read` scope, so "not permitted to see it" stays distinguishable
        from "no tourist tax applies".
      parameters:
        - name: apartmentId
          in: query
          required: true
          schema: { type: integer }
        - name: start
          in: query
          required: true
          description: Arrival date (Y-m-d).
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          description: Departure date (Y-m-d), exclusive. 1st to 8th is seven nights.
          schema: { type: string, format: date }
        - name: persons
          in: query
          description: >-
            Occupancy the apartment price is matched against. Required unless
            `guestCounts` carries occupancy-counted guests, in which case it defaults
            to their sum.
          schema: { type: integer, minimum: 1 }
        - name: guestCounts
          in: query
          style: deepObject
          explode: true
          description: >-
            Head count per guest category, as `guestCounts[categoryId]=count`. Drives
            the guest-category modifiers (child rates, extra-bed surcharges) and the
            tourist tax. Omit for a plain list price.
          schema:
            type: object
            additionalProperties: { type: integer, minimum: 0 }
        - $ref: '#/components/parameters/OriginId'
      responses:
        '200':
          description: The quote. The stay must not exceed 366 nights.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/PriceQuote' }
                  meta:
                    type: object
                    properties:
                      touristTaxIncluded: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /prices/rates:
    get:
      tags: [Prices]
      operationId: getRateCalendar
      summary: Per-night, per-occupancy rate calendar
      description: |
        Requires the `prices:read` scope.

        The rate for a night is not a property of that night alone: candidate prices are
        filtered by `minStay` and matched on occupancy exactly. Both are parameters here
        rather than assumptions, so the per-night rate of a seven-night stay
        (`nights=7`) can legitimately differ from the single-night rate (`nights=1`).

        A night with no applicable price comes back with an empty `rates` array.
      parameters:
        - name: roomCategoryId
          in: query
          description: Required unless `apartmentId` is given.
          schema: { type: integer }
        - name: apartmentId
          in: query
          description: Alternative to `roomCategoryId` — the room's category is used.
          schema: { type: integer }
        - name: start
          in: query
          description: First night (Y-m-d). Defaults to today.
          schema: { type: string, format: date }
        - name: end
          in: query
          description: Last night (Y-m-d), inclusive. Defaults to `start`.
          schema: { type: string, format: date }
        - name: nights
          in: query
          description: Intended stay length, drives minStay selection.
          schema: { type: integer, minimum: 1, default: 1 }
        - name: occupancy
          in: query
          description: >-
            Single occupancy to report. Defaults to every occupancy the category has an
            active price for.
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/OriginId'
      responses:
        '200':
          description: One entry per night. The range must not exceed 500 days.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        rates:
                          type: array
                          items: { $ref: '#/components/schemas/Rate' }
                  meta:
                    type: object
                    properties:
                      start: { type: string }
                      end: { type: string }
                      nights: { type: integer }
                      roomCategoryId: { type: [integer, 'null'] }
                      occupancies: { type: array, items: { type: integer } }
                      origin:
                        type: object
                        properties:
                          id: { type: integer }
                          name: { type: [string, 'null'] }
                      count: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Send the personal access token: `Authorization: Bearer fwb_...`'
    basicAuth:
      type: http
      scheme: basic
      description: >-
        Username of the token owner plus the personal access token as the
        password. Intended for clients that cannot send custom headers, such as
        calendar applications. The real login password is never accepted.
  parameters:
    OriginId:
      name: originId
      in: query
      description: >-
        Reservation origin the price rows are looked up for. Price rows are linked to
        origins, so different origins legitimately yield different prices. Defaults to
        the origin configured for online booking; when none is configured the parameter
        is required.
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid parameters.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid token.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            examples: ['Bearer realm="fewohbee-api", Basic realm="fewohbee-api"']
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The token lacks the required scope, or its owner lacks the role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Rate limit exceeded (300 requests per 5 minutes per token).
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: integer }
            message: { type: string }
      examples:
        - error:
            code: 400
            message: "Invalid parameter 'start': expected format Y-m-d."
...
    Price:
      type: object
      description: A configured price row. Reference data — see the endpoint description.
      properties:
        id: { type: integer }
        type: { type: string, enum: [apartment, misc] }
        description: { type: [string, 'null'] }
        price:
          type: number
          description: The stored amount. Per head, per room or for the whole stay depending on the flags below.
        vat: { type: number, description: 'VAT percentage.' }
        includesVat: { type: boolean, description: 'Whether `price` is gross.' }
        isFlatPrice: { type: boolean, description: 'Billed once for the stay.' }
        isPerRoom: { type: boolean, description: 'Billed per room-night, occupancy-independent.' }
        numberOfPersons:
          type: [integer, 'null']
          description: Occupancy this row is for. Matched exactly — there is no fallback to a lower occupancy.
        minStay: { type: [integer, 'null'] }
        active: { type: boolean }
        roomCategory:
          type: [object, 'null']
          properties:
            id: { type: [integer, 'null'] }
            name: { type: [string, 'null'] }
        origins:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: [string, 'null'] }
        allDays: { type: boolean, description: 'When true the weekday flags are ignored.' }
        weekdays:
          type: object
          description: Which weekdays the row applies to when `allDays` is false.
          properties:
            monday: { type: boolean }
            tuesday: { type: boolean }
            wednesday: { type: boolean }
            thursday: { type: boolean }
            friday: { type: boolean }
            saturday: { type: boolean }
            sunday: { type: boolean }
        allPeriods: { type: boolean, description: 'When true the row applies year-round and `periods` is irrelevant.' }
        periods:
          type: array
          description: Season windows the row applies to. Earlier-starting periods win over later ones.
          items:
            type: object
            properties:
              startDate: { type: string, format: date }
              endDate: { type: string, format: date }
        seasonStart: { type: [string, 'null'], format: date }
        seasonEnd: { type: [string, 'null'], format: date }
        isBookableOnline: { type: boolean }
        isMandatoryOnline: { type: boolean }
        isDefaultActiveInReservationCreation: { type: boolean }
        isPackage: { type: boolean, description: 'Whether the row splits into components for revenue accounting.' }
        components:
          type: array
          items:
            type: object
            properties:
              id: { type: [integer, 'null'] }
              description: { type: string }
              vat: { type: number }
              allocationType: { type: string }
              allocationValue: { type: [number, 'null'] }
              isRemainder: { type: boolean }
              sortOrder: { type: integer }
    Rate:
      type: object
      properties:
        occupancy: { type: integer }
        priceId: { type: integer }
        description: { type: [string, 'null'] }
        pricingModel: { type: string, enum: [per_person_night, per_room_night, flat] }
        unitPrice: { type: number, description: 'The stored amount of the price row.' }
        perNight:
          type: [number, 'null']
          description: >-
            Room total for one night at this occupancy. Null for `flat` rows — a flat
            price covers the whole stay and has no meaningful per-night figure.
        stayPrice:
          type: [number, 'null']
          description: The stay amount for `flat` rows; null otherwise.
        minStay: { type: [integer, 'null'] }
        vat: { type: number }
        includesVat: { type: boolean }
        isBookableOnline: { type: boolean }
...